Fixing docs.rs compilation error.
At the moment docs.rs fails to compile due to
not being able to link to Cuda or other backend
specific configurations, which it seemingly has
to do to compile the docs of those crates.

I have now removed the backends from being part
of the included features to be displayed on
docs.rs, so that only the crates that pertains
to the upper-level api remains.

As it was me who added the all-features flag in
the first place, I must apologize for any
inconvenience caused!